% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{add_annotation_property_value}
\alias{add_annotation_property_value}
\title{FUNCTION_TITLE}
\usage{
add_annotation_property_value(rdf, entity, annotation_property, value)
}
\arguments{
\item{rdf}{PARAM_DESCRIPTION}

\item{entity}{PARAM_DESCRIPTION}

\item{annotation_property}{PARAM_DESCRIPTION}

\item{value}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
\code{\link[rdflib]{rdf_add}}
}
