% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{add_class}
\alias{add_class}
\title{FUNCTION_TITLE}
\usage{
add_class(rdf, class_id, class_label)
}
\arguments{
\item{rdf}{PARAM_DESCRIPTION}

\item{class_id}{PARAM_DESCRIPTION}

\item{class_label}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rdflib]{rdf_add}}
}
