% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{add_individual_class}
\alias{add_individual_class}
\title{FUNCTION_TITLE}
\usage{
add_individual_class(rdf, individual, class)
}
\arguments{
\item{rdf}{PARAM_DESCRIPTION}

\item{individual}{PARAM_DESCRIPTION}

\item{class}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
\code{\link[rdflib]{rdf_add}}
}
