% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{add_subclassof}
\alias{add_subclassof}
\title{FUNCTION_TITLE}
\usage{
add_subclassof(rdf, parent_class, child_class)
}
\arguments{
\item{rdf}{PARAM_DESCRIPTION}

\item{parent_class}{PARAM_DESCRIPTION}

\item{child_class}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
\code{\link[rdflib]{rdf_add}}
}
