% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{append_annotation_properties}
\alias{append_annotation_properties}
\title{Append AnnotationProperty}
\usage{
append_annotation_properties(rdf, base_uri, ...)
}
\arguments{
\item{rdf}{PARAM_DESCRIPTION}

\item{base_uri}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
\verb{append_*} functions are to add an extension to the
RDF structure (i.e. addition of Object Properties, Annotation Properties, and
Data Properties) whereas \verb{add_*} and \verb{df_add_*} functions populate the RDF with
triplets.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{list2}}
\code{\link[glue]{glue}}
\code{\link[rdflib]{rdf_add}}
}
