% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{df_add_annotation_property_value}
\alias{df_add_annotation_property_value}
\title{FUNCTION_TITLE}
\usage{
df_add_annotation_property_value(
  rdf,
  data,
  entity_col,
  annotation_property_label,
  value_col
)
}
\arguments{
\item{rdf}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION}

\item{entity_col}{PARAM_DESCRIPTION}

\item{annotation_property_label}{PARAM_DESCRIPTION}

\item{value_col}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[cli]{cli_alert}}
\code{\link[dplyr]{filter}},\code{\link[dplyr]{select}}
}
