% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{df_add_individual_class}
\alias{df_add_individual_class}
\title{FUNCTION_TITLE}
\usage{
df_add_individual_class(rdf, data, class_id_col, individual_id_col)
}
\arguments{
\item{rdf}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION}

\item{class_id_col}{PARAM_DESCRIPTION}

\item{individual_id_col}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{select}}
\code{\link[rdflib]{rdf_add}}
}
