% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{df_add_subclassof}
\alias{df_add_subclassof}
\title{FUNCTION_TITLE}
\usage{
df_add_subclassof(rdf, data, parent_class_id_col, child_class_id_col)
}
\arguments{
\item{rdf}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION}

\item{parent_class_id_col}{PARAM_DESCRIPTION}

\item{child_class_id_col}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}
\code{\link[rdflib]{rdf_add}}
}
