% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{id_if_label}
\alias{id_if_label}
\title{FUNCTION_TITLE}
\usage{
id_if_label(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
