% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{initialize_blank_rdf}
\alias{initialize_blank_rdf}
\title{FUNCTION_TITLE}
\usage{
initialize_blank_rdf()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rdflib]{rdf_add}}
}
