% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{initialize_rdf}
\alias{initialize_rdf}
\title{FUNCTION_TITLE}
\usage{
initialize_rdf(name, base_uri, version, comment)
}
\arguments{
\item{name}{PARAM_DESCRIPTION}

\item{base_uri}{PARAM_DESCRIPTION}

\item{version}{PARAM_DESCRIPTION}

\item{comment}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rdflib]{rdf_add}}
\code{\link[glue]{glue}}
}
