% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{is_empty_document}
\alias{is_empty_document}
\title{Is the RDF File Empty?}
\usage{
is_empty_document(
  doc,
  format = c("rdfxml", "nquads", "ntriples", "turtle", "jsonld")
)
}
\arguments{
\item{doc}{PARAM_DESCRIPTION}

\item{format}{PARAM_DESCRIPTION, Default: c("rdfxml", "nquads", "ntriples", "turtle", "jsonld")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Check if a given file has 1 or more triplets.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
