% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{label_to_id}
\alias{label_to_id}
\title{FUNCTION_TITLE}
\usage{
label_to_id(x, rdf)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{rdf}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{select}}
\code{\link[cli]{cli_alert}}
}
