% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql_wrapper.R
\name{map_to_labels}
\alias{map_to_labels}
\title{FUNCTION_TITLE}
\usage{
map_to_labels(data, ..., rdf)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{rdf}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}}
\code{\link[rlang]{parse_expr}}
}
