% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{open_in_protege}
\alias{open_in_protege}
\title{FUNCTION_TITLE}
\usage{
open_in_protege(doc)
}
\arguments{
\item{doc}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
