% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql_wrapper.R
\name{query_annotation_predicates}
\alias{query_annotation_predicates}
\title{FUNCTION_TITLE}
\usage{
query_annotation_predicates(rdf)
}
\arguments{
\item{rdf}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rdflib]{rdf_query}}
\code{\link[dplyr]{mutate-joins}},\code{\link[dplyr]{select}}
\code{\link[purrr]{map}}
}
