% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql_wrapper.R
\name{query_rdf_taxonomy}
\alias{query_rdf_taxonomy}
\title{FUNCTION_TITLE}
\usage{
query_rdf_taxonomy(rdf)
}
\arguments{
\item{rdf}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rdflib]{rdf_query}}
\code{\link[dplyr]{filter}},\code{\link[dplyr]{distinct}},\code{\link[dplyr]{mutate-joins}}
}
