% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql_wrapper.R
\name{query_type_map}
\alias{query_type_map}
\title{FUNCTION_TITLE}
\usage{
query_type_map(rdf)
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rdflib]{rdf_query}}
}
