% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql_wrapper.R
\name{rdf_as_df}
\alias{rdf_as_df}
\title{FUNCTION_TITLE}
\usage{
rdf_as_df(rdf)
}
\arguments{
\item{rdf}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
