% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql_wrapper.R
\name{read_label_map}
\alias{read_label_map}
\title{FUNCTION_TITLE}
\usage{
read_label_map(doc)
}
\arguments{
\item{doc}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rdflib]{rdf_query}}
}
