% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql_wrapper.R
\name{read_pivoted_rdf_taxonomy}
\alias{read_pivoted_rdf_taxonomy}
\title{FUNCTION_TITLE}
\usage{
read_pivoted_rdf_taxonomy(doc)
}
\arguments{
\item{doc}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate-joins}},\code{\link[dplyr]{rename}},\code{\link[dplyr]{tidyeval-compat}}
\code{\link[glue]{glue}}
}
