% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{read_rdf}
\alias{read_rdf}
\title{FUNCTION_TITLE}
\usage{
read_rdf(doc, formats = c("rdfxml", "turtle"))
}
\arguments{
\item{doc}{PARAM_DESCRIPTION}

\item{formats}{PARAM_DESCRIPTION, Default: c("rdfxml", "turtle")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{safely}},\code{\link[purrr]{map}}
\code{\link[tibble]{tibble}}
\code{\link[utils]{capture.output}}
\code{\link[huxtable]{huxtable}},\code{\link[huxtable]{themes}},\code{\link[huxtable]{na_string}},\code{\link[huxtable]{number_format}},\code{\link[huxtable]{fmt_pretty}}
\code{\link[cli]{character(0)}},\code{\link[cli]{cli_alert}}
\code{\link[rdflib]{rdf_parse}}
}
