% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{read_rdf2}
\alias{read_rdf2}
\title{FUNCTION_TITLE}
\usage{
read_rdf2(doc)
}
\arguments{
\item{doc}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[xfun]{file_ext}}
\code{\link[rdflib]{rdf_parse}}
\code{\link[cli]{cli_alert}}
}
