% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{read_rdf_fmts}
\alias{read_rdf_fmts}
\title{FUNCTION_TITLE}
\usage{
read_rdf_fmts(doc, formats = c("rdfxml", "turtle"))
}
\arguments{
\item{doc}{PARAM_DESCRIPTION}

\item{formats}{PARAM_DESCRIPTION, Default: c("rdfxml", "turtle")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rdflib]{rdf_parse}}
\code{\link[cli]{cli_alert}}
}
