% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql_wrapper.R
\name{search_by_predicate}
\alias{search_by_predicate}
\title{FUNCTION_TITLE}
\usage{
search_by_predicate(predicate, rdf_or_doc)
}
\arguments{
\item{predicate}{PARAM_DESCRIPTION}

\item{rdf_or_doc}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rdflib]{rdf_query}}
}
