% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{uri_format}
\alias{uri_format}
\title{Format a URI}
\usage{
uri_format(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Strict formatting guidelines are applied to any URI to prevent errors at the R-Protege interface.
Replaces any punctuation of length one or more with a dash and only keeps characters with alphanumeric characters or a dash.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_replace}},\code{\link[stringr]{str_remove}}
}
