% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{write_rdf}
\alias{write_rdf}
\title{Write RDF}
\usage{
write_rdf(rdf, doc)
}
\description{
Writes RDF exclusively as a Turtle file because it is the only format that
Protege properly loads when written via R.
}
\seealso{
\code{\link[cli]{character(0)}}
\code{\link[rdflib]{rdf_serialize}}
}
